requirements.txt
PyQt6>=6.4.0
GitPython>=3.1.0
